place order service and get orders services developed

+++ seller/src/main/java/com/sulek/seller/dto/ProductResponseDto.java
@@ -4,10 +4,12 @@ package com.sulek.seller.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ProductResponseDto {
-    private Long quantity;
+    private BigDecimal quantity;
     private String name;
     private String description;
 }
